chore: update dependencies and fix related issues for upgrade
Former-commit-id: 0e447a85b266dbd3b0fdc4c4afaa1fa77407d08b
Former-commit-id: 5ae939080aa8c47492f58837a4a8fd7fce5c28d8
Former-commit-id: 10cc09928a02eb976ec9d8b3be66301878b1d6a0

+++ src/fasta_reader.cpp
@@ -32,7 +32,7 @@ class FastaReader::Impl {
     const auto regionString = region.ToRegionString();
     char* rawSeqData = fai_fetch(idx.get(), regionString.c_str(), &parsedSeqLen);
 
-    if (rawSeqData == nullptr || parsedSeqLen != expectedSeqLen) {
+    if (rawSeqData == nullptr || static_cast<std::size_t>(parsedSeqLen) != expectedSeqLen) {
       const auto errMsg = absl::StrFormat("could not fetch sequence for %s from %s", regionString, srcPath);
       return rawSeqData == nullptr ? absl::InternalError(errMsg) : absl::FailedPreconditionError(errMsg);
     }
